DBU-652: split Pyth adapter into modules
